pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "An e-commerce website project for learning"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-htmx>=1.27.0",
    "pillow>=12.1.0",
]
